dataset/cache/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
requests==2.31.0
psycopg2-binary==2.9.9